+++ HISTORY.md
@@ -1,6 +1,6 @@
 unreleased
 ==========
-  * Add the defaultEncoding option for requests without `Accept-Encoding` header
+  * Add the enforceEncoding option for requests without `Accept-Encoding` header
 
 1.7.5 / 2024-10-31
 ==========